seperate xml utils
